Removed cereal from header files by forward declaration.
