+++ src/openMVG/features/io_regions_type.hpp
@@ -9,41 +9,16 @@
 #ifndef OPENMVG_FEATURES_IO_REGIONS_TYPE_HPP
 #define OPENMVG_FEATURES_IO_REGIONS_TYPE_HPP
 
-#include <fstream>
-#include <string>
-#include <vector>
-
-#include "openMVG/features/regions_factory.hpp"
-#include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
-
-#include <cereal/archives/json.hpp>
+#include "openMVG/features/regions.hpp"
 
 namespace openMVG {
 namespace features {
 
 // Init the regions_type from an image describer file (used for regions loading)
-inline std::unique_ptr<features::Regions> Init_region_type_from_file
+std::unique_ptr<features::Regions> Init_region_type_from_file
 (
   const std::string & sImage_describer_file
-)
-{
-  std::unique_ptr<Regions> regions_type;
-  if (stlplus::is_file(sImage_describer_file))
-  {
-    // Dynamically load the regions type from the file
-    std::ifstream stream(sImage_describer_file.c_str());
-    if (stream.is_open())
-    {
-      cereal::JSONInputArchive archive(stream);
-      archive(cereal::make_nvp("regions_type", regions_type));
-    }
-  }
-  else // By default init a SIFT regions type (keep compatibility)
-  {
-    regions_type.reset(new features::SIFT_Regions());
-  }
-  return regions_type;
-}
+);
 
 } // namespace features
 } // namespace openMVG
