+++ src/openMVG/features/sift/SIFT_Anatomy_Image_Describer.hpp
@@ -48,19 +48,8 @@ Changes are:
 #ifndef OPENMVG_FEATURES_SIFT_SIFT_ANATOMY_IMAGE_DESCRIBER_HPP
 #define OPENMVG_FEATURES_SIFT_SIFT_ANATOMY_IMAGE_DESCRIBER_HPP
 
-#include <iostream>
-#include <numeric>
-#include <vector>
 
-#include "openMVG/features/feature.hpp"
 #include "openMVG/features/image_describer.hpp"
-#include "openMVG/features/regions_factory.hpp"
-#include "openMVG/features/sift/hierarchical_gaussian_scale_space.hpp"
-#include "openMVG/features/sift/sift_DescriptorExtractor.hpp"
-#include "openMVG/features/sift/sift_keypoint.hpp"
-#include "openMVG/features/sift/sift_KeypointExtractor.hpp"
-
-#include <cereal/cereal.hpp>
 
 namespace openMVG {
 namespace features {
@@ -77,25 +66,10 @@ public:
       float edge_threshold = 10.0f,
       float peak_threshold = 0.04f,
       bool root_sift = true
-    ):
-      first_octave_(first_octave),
-      num_octaves_(num_octaves),
-      num_scales_(num_scales),
-      edge_threshold_(edge_threshold),
-      peak_threshold_(peak_threshold),
-      root_sift_(root_sift) {}
+    );
 
     template<class Archive>
-    void serialize( Archive & ar )
-    {
-      ar(
-        cereal::make_nvp("first_octave", first_octave_),
-        cereal::make_nvp("num_octaves",num_octaves_),
-        cereal::make_nvp("num_scales",num_scales_),
-        cereal::make_nvp("edge_threshold",edge_threshold_),
-        cereal::make_nvp("peak_threshold",peak_threshold_),
-        cereal::make_nvp("root_sift",root_sift_));
-    }
+    void serialize( Archive & ar );
 
     // Parameters
     int first_octave_;      // Use original image, or perform an upscale if == -1
@@ -109,29 +83,9 @@ public:
   SIFT_Anatomy_Image_describer
   (
     const Params params = Params()
-  )
-  :Image_describer(), params_(params)
-  {}
+  );
 
-  bool Set_configuration_preset(EDESCRIBER_PRESET preset) override
-  {
-    switch(preset)
-    {
-    case NORMAL_PRESET:
-      params_.peak_threshold_ = 0.04f;
-    break;
-    case HIGH_PRESET:
-      params_.peak_threshold_ = 0.01f;
-    break;
-    case ULTRA_PRESET:
-      params_.peak_threshold_ = 0.01f;
-      params_.first_octave_ = -1;
-    break;
-    default:
-      return false;
-    }
-    return true;
-  }
+  bool Set_configuration_preset(EDESCRIBER_PRESET preset) override;
 
   /**
   @brief Detect regions on the image and compute their attributes (description)
@@ -145,81 +99,13 @@ public:
     const image::Image<unsigned char>& image,
     std::unique_ptr<Regions> &regions,
     const image::Image<unsigned char> * mask = nullptr
-  ) override
-  {
-    // Convert to float in range [0;1]
-    const image::Image<float> If(image.GetMat().cast<float>()/255.0f);
-
-    // compute sift keypoints
-    Allocate(regions);
-
-    // Build alias to cached data
-    SIFT_Regions * regionsCasted = dynamic_cast<SIFT_Regions*>(regions.get());
-    {
-      using namespace openMVG::features::sift;
-      const int supplementary_images = 3;
-      // => in order to ensure each gaussian slice is used in the process 3 extra images are required:
-      // +1 for dog computation
-      // +2 for 3d discrete extrema definition
-
-      HierarchicalGaussianScaleSpace octave_gen(
-        params_.num_octaves_,
-        params_.num_scales_,
-        (params_.first_octave_ == -1)
-        ? GaussianScaleSpaceParams(1.6f/2.0f, 1.0f/2.0f, 0.5f, supplementary_images)
-        : GaussianScaleSpaceParams(1.6f, 1.0f, 0.5f, supplementary_images));
-      octave_gen.SetImage( If );
-
-      std::vector<Keypoint> keypoints;
-      keypoints.reserve(5000);
-      Octave octave;
-      while ( octave_gen.NextOctave( octave ) )
-      {
-        std::vector< Keypoint > keys;
-        // Find Keypoints
-        SIFT_KeypointExtractor keypointDetector(
-          params_.peak_threshold_ / octave_gen.NbSlice(),
-          params_.edge_threshold_);
-        keypointDetector(octave, keys);
-        // Find Keypoints orientation and compute their description
-        Sift_DescriptorExtractor descriptorExtractor;
-        descriptorExtractor(octave, keys);
-
-        // Concatenate the found keypoints
-        std::move(keys.begin(), keys.end(), std::back_inserter(keypoints));
-      }
-      for (const auto & k : keypoints)
-      {
-        // Feature masking
-        if (mask)
-        {
-          const image::Image<unsigned char> & maskIma = *mask;
-          if (maskIma(k.y, k.x) == 0)
-            continue;
-        }
-
-        Descriptor<unsigned char, 128> descriptor;
-        descriptor << (k.descr.cast<unsigned char>());
-        {
-          regionsCasted->Descriptors().emplace_back(descriptor);
-          regionsCasted->Features().emplace_back(k.x, k.y, k.sigma, k.theta);
-        }
-      }
-    }
-    return true;
-  };
+  ) override;
 
   /// Allocate Regions type depending of the Image_describer
-  void Allocate(std::unique_ptr<Regions> &regions) const override
-  {
-    regions.reset( new SIFT_Regions );
-  }
+  void Allocate(std::unique_ptr<Regions> &regions) const override;
 
   template<class Archive>
-  void serialize( Archive & ar )
-  {
-    ar(cereal::make_nvp("params", params_));
-  }
+  void serialize( Archive & ar );
 
 private:
   Params params_;
@@ -227,10 +113,5 @@ private:
 
 } // namespace features
 } // namespace openMVG
-
-#include <cereal/types/polymorphic.hpp>
-#include <cereal/archives/json.hpp>
-CEREAL_REGISTER_TYPE_WITH_NAME(openMVG::features::SIFT_Anatomy_Image_describer, "SIFT_Anatomy_Image_describer");
-CEREAL_REGISTER_POLYMORPHIC_RELATION(openMVG::features::Image_describer, openMVG::features::SIFT_Anatomy_Image_describer)
 
 #endif // OPENMVG_FEATURES_SIFT_SIFT_ANATOMY_IMAGE_DESCRIBER_HPP
