+++ src/openMVG/sfm/pipelines/sfm_features_provider.hpp
@@ -19,6 +19,7 @@
 #include "openMVG/types.hpp"
 
 #include "third_party/progress/progress_display.hpp"
+#include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
 
 namespace openMVG {
 namespace sfm {
