+++ src/openMVG/sfm/sfm_data_io_cereal.cpp
@@ -11,6 +11,7 @@
 #include "openMVG/cameras/cameras.hpp"
 #include "openMVG/sfm/sfm_data.hpp"
 #include "openMVG/types.hpp"
+#include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
 
 #include <fstream>
 #include <string>
