+++ src/openMVG_Samples/features_kvld_filter/kvld_filter.cpp
@@ -7,6 +7,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #include "openMVG/features/svg_features.hpp"
+#include "openMVG/features/regions_factory.hpp"
 #include "openMVG/image/image_io.hpp"
 #include "openMVG/image/image_concat.hpp"
 #include "openMVG/matching/kvld/kvld.h"
