+++ src/openMVG_Samples/image_undistort_gui/CMakeLists.txt
@@ -10,7 +10,7 @@ if( Qt5Widgets_FOUND )
   set( OPENMVG_SAMPLE_IMAGE_UNDISTORT_HDRS MainWindow.hh QImageInterface.hh )
 
   add_executable( openMVG_sample_image_undistort ${OPENMVG_SAMPLE_IMAGE_UNDISTORT_SRCS} ${OPENMVG_SAMPLE_IMAGE_UNDISTORT_HDRS})
-  target_link_libraries( openMVG_sample_image_undistort Qt5::Widgets openMVG_image openMVG_multiview )
+  target_link_libraries( openMVG_sample_image_undistort Qt5::Widgets openMVG_image openMVG_multiview openMVG_geometry )
 
-   set_property(TARGET openMVG_sample_image_undistort PROPERTY CXX_STANDARD 11)
+  set_property(TARGET openMVG_sample_image_undistort PROPERTY CXX_STANDARD 11)
 endif( Qt5Widgets_FOUND)
