+++ src/openMVG_Samples/multiview_robust_essential_spherical/robust_essential_spherical.cpp
@@ -9,6 +9,7 @@
 
 #include "openMVG/cameras/Camera_Spherical.hpp"
 #include "openMVG/features/feature.hpp"
+#include "openMVG/features/regions_factory.hpp"
 #include "openMVG/features/svg_features.hpp"
 #include "openMVG/image/image_io.hpp"
 #include "openMVG/image/image_concat.hpp"
