+++ src/software/SfM/main_ExportUndistortedImages.cpp
@@ -14,6 +14,7 @@
 
 #include "third_party/cmdLine/cmdLine.h"
 #include "third_party/progress/progress_display.hpp"
+#include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
 
 #include <cstdlib>
 #include <string>
