+++ src/software/SfM/main_openMVG2CMPMVS.cpp
@@ -19,6 +19,7 @@ using namespace openMVG::sfm;
 
 #include "third_party/cmdLine/cmdLine.h"
 #include "third_party/progress/progress_display.hpp"
+#include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
 
 #include <cstdlib>
 #include <cmath>
