+++ src/software/SfMWebGLViewer/openMVG_main_openMVG2WebGL.cpp
@@ -18,6 +18,7 @@
 
 #include "third_party/cmdLine/cmdLine.h"
 #include "third_party/progress/progress_display.hpp"
+#include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
 
 #include <iostream>
 #include <fstream>
